Made Interop.UIAutomationCore.dll platform-agnostic (AnyCPU).
It was not enough to set /machine:Agnostic in command-line arguments of tlbimp because for some reason it then sets interop structure packing to 4 which works only on x86 really. After some investigation I found that there's 5 structures in the Interop.UIAutomationCore.dll where x86 and x64 versions of tlbimp produced different packing (4 on x86 and 8 on x64). All of these places were changed to produce platform-default packing (".pack 0" instruction in the IL) which produces exactly 4 on x86 and 8 on x64. Tests are passing on both platforms successfully.